+++ common/src/main/kotlin/com/cobblemon/mod/common/client/particle/ParticleStorm.kt
@@ -9,16 +9,21 @@
 package com.cobblemon.mod.common.client.particle
 
 import com.bedrockk.molang.runtime.MoLangRuntime
-import com.bedrockk.molang.runtime.struct.VariableStruct
 import com.bedrockk.molang.runtime.value.DoubleValue
 import com.cobblemon.mod.common.api.snowstorm.BedrockParticleEffect
 import com.cobblemon.mod.common.api.snowstorm.ParticleEmitterAction
+import com.cobblemon.mod.common.client.render.MatrixWrapper
 import com.cobblemon.mod.common.client.render.SnowstormParticle
 import com.cobblemon.mod.common.particle.SnowstormParticleEffect
+import com.cobblemon.mod.common.util.math.geometry.getOrigin
+import com.cobblemon.mod.common.util.math.geometry.transformDirection
+import com.cobblemon.mod.common.util.math.geometry.transformPosition
 import kotlin.random.Random
+import net.minecraft.client.MinecraftClient
 import net.minecraft.client.particle.NoRenderParticle
 import net.minecraft.client.world.ClientWorld
 import net.minecraft.entity.Entity
+import net.minecraft.util.math.Vec3d
 
 /**
  * An instance of a bedrock particle effect.
@@ -28,12 +33,21 @@ import net.minecraft.entity.Entity
  */
 class ParticleStorm(
     val effect: BedrockParticleEffect,
-    val origin: SnowstormParticleOrigin,
-    val world: ClientWorld
-): NoRenderParticle(world, origin.position.x, origin.position.y, origin.position.z) {
-    val runtime = MoLangRuntime().also {
-        it.environment.structs["variable"] = VariableStruct()
+    private val matrixWrapper: MatrixWrapper,
+    val world: ClientWorld,
+    val sourceVelocity: () -> Vec3d = { Vec3d.ZERO },
+    val runtime: MoLangRuntime = MoLangRuntime()
+): NoRenderParticle(world, matrixWrapper.matrix.getOrigin().x, matrixWrapper.matrix.getOrigin().y, matrixWrapper.matrix.getOrigin().z) {
+    fun spawn() {
+        MinecraftClient.getInstance().particleManager.addParticle(this)
     }
+    fun getX() = x
+    fun getY() = y
+    fun getZ() = z
+
+    fun getPrevX() = prevPosX
+    fun getPrevY() = prevPosY
+    fun getPrevZ() = prevPosZ
 
     val particles = mutableListOf<SnowstormParticle>()
     var started = false
@@ -62,7 +76,11 @@ class ParticleStorm(
         setMaxAge(getMaxAge())
         super.tick()
 
-        val pos = origin.position
+        val pos = matrixWrapper.matrix.getOrigin()
+        prevPosX = x
+        prevPosY = y
+        prevPosZ = z
+
         x = pos.x
         y = pos.y
         z = pos.z
@@ -77,8 +95,8 @@ class ParticleStorm(
 
         when (effect.emitter.lifetime.getAction(runtime, started, age / 20.0)) {
             ParticleEmitterAction.GO -> {
+                val toEmit = effect.emitter.rate.getEmitCount(runtime, started, particles.size)
                 started = true
-                val toEmit = effect.emitter.rate.getEmitCount(runtime, particles.size)
                 repeat(times = toEmit) {
                     spawnParticle()
                 }
@@ -95,10 +113,16 @@ class ParticleStorm(
         runtime.environment.setSimpleVariable("particle_random_3", DoubleValue(Random.Default.nextDouble()))
         runtime.environment.setSimpleVariable("particle_random_4", DoubleValue(Random.Default.nextDouble()))
 
-        val center = effect.emitter.shape.getCenter(runtime, entity).add(x, y, z)
-        val newPosition = effect.emitter.shape.getNewParticlePosition(runtime, entity).add(x, y, z)
-        val velocity = effect.particle.motion.getInitialVelocity(runtime, particlePos = newPosition, emitterPos = center).multiply(1/50.0)
-        velocity.multiply(1 / 20.0)
+        val center = transformPosition(effect.emitter.shape.getCenter(runtime, entity))
+        val newPosition = transformPosition(effect.emitter.shape.getNewParticlePosition(runtime, entity))
+        val initialVelocity = effect.particle.motion.getInitialVelocity(runtime, particlePos = newPosition, emitterPos = center)
+        val velocity = (if (effect.space.localPosition) transformDirection(initialVelocity) else initialVelocity)
+            .multiply(1 / 20.0)
+            .add(if (effect.space.localVelocity) sourceVelocity() else Vec3d.ZERO)
+
         world.addParticle(particleEffect, newPosition.x, newPosition.y, newPosition.z, velocity.x, velocity.y, velocity.z)
     }
+
+    fun transformPosition(position: Vec3d): Vec3d = matrixWrapper.matrix.transformPosition(position)
+    fun transformDirection(direction: Vec3d): Vec3d = matrixWrapper.matrix.transformDirection(direction)
 }