+++ common/src/main/kotlin/com/cobblemon/mod/common/client/render/models/blockbench/bedrock/animation/BedrockAnimationAdapter.kt
@@ -13,12 +13,16 @@ import com.bedrockk.molang.parser.MoLangParser
 import com.bedrockk.molang.parser.tokenizer.TokenIterator
 import com.bedrockk.molang.runtime.MoLangRuntime
 import com.cobblemon.mod.common.Cobblemon.LOGGER
+import com.cobblemon.mod.common.client.particle.BedrockParticleEffectRepository
+import com.cobblemon.mod.common.util.asIdentifierDefaultingNamespace
 import com.google.gson.JsonArray
 import com.google.gson.JsonDeserializationContext
 import com.google.gson.JsonDeserializer
 import com.google.gson.JsonElement
 import com.google.gson.JsonObject
+import java.lang.IllegalArgumentException
 import java.lang.reflect.Type
+import net.minecraft.util.Identifier
 
 /**
  * Gson adapter for converting bedrock/blockbench json data into a friendlier object model.
@@ -27,19 +31,35 @@ import java.lang.reflect.Type
  * @since  January 5, 2022
  */
 object BedrockAnimationAdapter : JsonDeserializer<BedrockAnimation> {
-    val molangRuntime = MoLangRuntime().also {
-        it.environment.structs["query"] = it.environment.structs["variable"]
-    }
-
     override fun deserialize(json: JsonElement, typeOfT: Type, context: JsonDeserializationContext): BedrockAnimation {
         if (json is JsonObject) {
             val animationLength = json["animation_length"]?.asDouble ?: -1.0
             val shouldLoop = animationLength > 0 && json["loop"]?.asBoolean == true
             val boneTimelines = mutableMapOf<String, BedrockBoneTimeline>()
+            val particleEffects = mutableListOf<BedrockParticleKeyframe>()
             json["bones"].asJsonObject.entrySet().forEach { (boneName, timeline) ->
                 boneTimelines[boneName] = deserializeBoneTimeline(timeline.asJsonObject)
             }
-            return BedrockAnimation(shouldLoop, animationLength, boneTimelines)
+            json["particle_effects"]?.asJsonObject?.entrySet()?.forEach { (frame, effectJson) ->
+                effectJson as JsonObject
+                val effectId = effectJson.get("effect").asString.asIdentifierDefaultingNamespace()
+                val effect = BedrockParticleEffectRepository.getEffect(effectId)
+                    ?: throw IllegalArgumentException("Unrecognized particle effect $effectId referenced in animation. Maybe your particle effect isn't named correctly inside the effect file?")
+                val locator = effectJson.get("locator")?.asString ?: "root"
+                val seconds = frame.toFloat()
+                val scripts = effectJson.get("pre_effect_script")?.asString?.split("\n")?.map { MoLang.createParser(it).parseExpression() } ?: emptyList()
+
+                particleEffects.add(
+                    BedrockParticleKeyframe(
+                        seconds = seconds,
+                        effect = effect,
+                        locator = locator,
+                        scripts = scripts
+                    )
+                )
+            }
+            // TODO it val effects =
+            return BedrockAnimation(shouldLoop, animationLength, particleEffects, boneTimelines)
         }
         else {
             throw IllegalStateException("animation json could not be parsed")